:art: move definitions models to own module
